apps/api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
imapclient
aioimaplib
aiosmtplib
zeroconf
websockets
python-multipart
requests
python-jose[cryptography]
cryptography
passlib[bcrypt]
alembic
